requirements.txt
Django==1.5.3
South==0.8.2
distribute==0.6.34
dj-database-url==0.2.2
dj-static==0.0.5
django-tastypie==0.11.0
django-toolbelt==0.0.1
gunicorn==18.0
markdown2==2.2.1
psycopg2==2.5.1
pystache==0.5.3
python-dateutil==2.2
python-mimeparse==0.1.4
requests==2.2.1
simplejson==3.3.2
six==1.6.1
static==0.4
wsgiref==0.1.2
